+++ packages/react-charts/src/victory/components/ChartBullet/examples/ChartBulletNegativeMeasure.tsx
@@ -1,21 +1,19 @@
-import { ChartBullet, ChartThemeColor } from '@patternfly/react-charts/victory';
+import { ChartBullet } from '@patternfly/react-charts/victory';
 
 interface ChartData {
   name: string;
   y?: number;
+  y0?: number;
 }
 
-export const ChartBulletPrimaryRange: React.FunctionComponent = () => {
-  const comparativeWarningMeasureData: ChartData[] = [{ name: 'Warning', y: 80 }];
+export const ChartBulletNegativeMeasure: React.FunctionComponent = () => {
+  const comparativeWarningMeasureData: ChartData[] = [{ name: 'Warning', y: 60 }];
   const comparativeWarningMeasureLegendData: ChartData[] = [{ name: 'Warning' }];
-  const primarySegmentedMeasureData: ChartData[] = [
-    { name: 'Measure', y: 75 },
-    { name: 'Measure', y: 135 }
-  ];
-  const primarySegmentedMeasureLegendData: ChartData[] = [{ name: 'Measure 1' }, { name: 'Measure 2' }];
+  const primarySegmentedMeasureData: ChartData[] = [{ name: 'Measure', y: -15 }];
+  const primarySegmentedMeasureLegendData: ChartData[] = [{ name: 'Measure 1' }];
   const qualitativeRangeData: ChartData[] = [
-    { name: 'Range', y: 85 },
-    { name: 'Range', y: 125 }
+    { name: 'Range', y: 25, y0: -25 },
+    { name: 'Range', y: 50 }
   ];
   const qualitativeRangeLegendData: ChartData[] = [{ name: 'Range 1' }, { name: 'Range 2' }];
 
@@ -27,22 +25,21 @@ export const ChartBulletPrimaryRange: React.FunctionComponent = () => {
         comparativeWarningMeasureData={comparativeWarningMeasureData}
         comparativeWarningMeasureLegendData={comparativeWarningMeasureLegendData}
         constrainToVisibleArea
-        labels={({ datum }) => `${datum.name}: ${datum.y}`}
         height={200}
-        maxDomain={{ y: 125 }}
-        minDomain={{ y: 50 }}
-        name="chart6"
+        labels={({ datum }) => `${datum.name}: ${datum.y}`}
+        maxDomain={{ y: 75 }}
+        minDomain={{ y: -25 }}
+        name="chart7"
         padding={{
           bottom: 50,
           left: 150, // Adjusted to accommodate labels
-          right: 75,
-          top: 50
+          right: 50,
+          top: 65
         }}
         primarySegmentedMeasureData={primarySegmentedMeasureData}
         primarySegmentedMeasureLegendData={primarySegmentedMeasureLegendData}
         qualitativeRangeData={qualitativeRangeData}
         qualitativeRangeLegendData={qualitativeRangeLegendData}
-        themeColor={ChartThemeColor.yellow}
         subTitle="Measure details"
         title="Text label"
         width={600}
