+++ packages/react-charts/src/victory/components/ChartBullet/examples/ChartBulletNegativePositiveMeasure.tsx
@@ -3,43 +3,49 @@ import { ChartBullet } from '@patternfly/react-charts/victory';
 interface ChartData {
   name: string;
   y?: number;
+  y0?: number;
 }
 
-export const ChartBulletReversedLegend: React.FunctionComponent = () => {
-  const comparativeWarningMeasureData: ChartData[] = [{ name: 'Warning', y: -88 }];
+export const ChartBulletNegativePositiveMeasure: React.FunctionComponent = () => {
+  const comparativeWarningMeasureData: ChartData[] = [{ name: 'Warning', y: 60 }];
   const comparativeWarningMeasureLegendData: ChartData[] = [{ name: 'Warning' }];
   const primarySegmentedMeasureData: ChartData[] = [
-    { name: 'Measure', y: -60 },
-    { name: 'Measure', y: -25 }
+    { name: 'Measure', y: -10 },
+    { name: 'Measure', y: -20 },
+    { name: 'Measure', y: 10 },
+    { name: 'Measure', y: 40 }
   ];
-  const primarySegmentedMeasureLegendData: ChartData[] = [{ name: 'Measure 1' }, { name: 'Measure 2' }];
+  const primarySegmentedMeasureLegendData: ChartData[] = [
+    { name: 'Measure 1' },
+    { name: 'Measure 2' },
+    { name: 'Measure 3' },
+    { name: 'Measure 4' }
+  ];
   const qualitativeRangeData: ChartData[] = [
-    { name: 'Range', y: -50 },
-    { name: 'Range', y: -75 }
+    { name: 'Range', y: 25, y0: -25 },
+    { name: 'Range', y: 50 }
   ];
   const qualitativeRangeLegendData: ChartData[] = [{ name: 'Range 1' }, { name: 'Range 2' }];
 
   return (
-    <div style={{ height: '200px', width: '700px' }}>
+    <div style={{ height: '200px', width: '600px' }}>
       <ChartBullet
         ariaDesc="Storage capacity"
         ariaTitle="Bullet chart example"
         comparativeWarningMeasureData={comparativeWarningMeasureData}
         comparativeWarningMeasureLegendData={comparativeWarningMeasureLegendData}
         constrainToVisibleArea
-        invert
         height={200}
         labels={({ datum }) => `${datum.name}: ${datum.y}`}
-        legendPosition="right"
-        legendOrientation="vertical"
-        maxDomain={{ y: 0 }}
-        minDomain={{ y: -100 }}
-        name="chart8"
+        legendItemsPerRow={4}
+        maxDomain={{ y: 75 }}
+        minDomain={{ y: -25 }}
+        name="chart9"
         padding={{
           bottom: 50,
           left: 150, // Adjusted to accommodate labels
-          right: 150, // Adjusted to accommodate legend
-          top: 80
+          right: 50,
+          top: 65
         }}
         primarySegmentedMeasureData={primarySegmentedMeasureData}
         primarySegmentedMeasureLegendData={primarySegmentedMeasureLegendData}
@@ -47,7 +53,7 @@ export const ChartBulletReversedLegend: React.FunctionComponent = () => {
         qualitativeRangeLegendData={qualitativeRangeLegendData}
         subTitle="Measure details"
         title="Text label"
-        width={700}
+        width={600}
       />
     </div>
   );
